The compiler totally fails to generate optimal output when the 128-bit integer wrapper package I wrote for the sake of readability.

```
name             old time/op  new time/op  delta
VerifyBatch64-8  2.74ms ± 0%  2.18ms ± 0%  -20.73%
KeyGeneration-8  23.8µs ± 0%  19.3µs ± 0%  -19.06%
Signing-8        24.5µs ± 0%  20.0µs ± 0%  -18.55%
Verification-8   78.6µs ± 0%  58.8µs ± 0%  -25.26%
```